chore: separate admin islands into a subfolder
