+++ src/islands/administration/index.ts
@@ -1,0 +1,7 @@
+import { initAdminTransfer } from './admin-transfer';
+import { initAdminLogo } from './admin-logo';
+
+export default (root: HTMLElement) => {
+    initAdminTransfer(root);
+    initAdminLogo(root);
+};
