+++ src/styles/admin.css
@@ -1,3 +1,10 @@
+.admin-panel {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    gap: 1rem;
+}
+
 #admin-sql-query-input {
     font-family: monospace;
     font-size: 14px;
@@ -54,6 +61,7 @@
     border: 1px solid var(--border);
     border-radius: 0.5rem;
     background: color-mix(in srgb, var(--bg) 92%, var(--fg) 4%);
+    flex: auto;
 }
 
 .admin-transfer h2 {
@@ -65,6 +73,41 @@
     flex-wrap: wrap;
     gap: 0.75rem;
     margin-top: 0.75rem;
+}
+
+.admin-logo {
+    margin-bottom: 1.5rem;
+    padding: 1rem;
+    border: 1px solid var(--border);
+    border-radius: 0.5rem;
+    background: color-mix(in srgb, var(--bg) 92%, var(--fg) 4%);
+    flex: auto;
+}
+
+.admin-logo h2 {
+    margin-top: 0;
+}
+
+.admin-logo-preview {
+    --pad: 0.5rem;
+    --pad-2x: calc(var(--pad) * 2);
+    width: calc(130px + var(--pad-2x));
+    height: calc(30px + var(--pad-2x));
+    padding: var(--pad);
+    display: grid;
+    place-items: center;
+    border: 1px dashed var(--border);
+    border-radius: 999px;
+    overflow: hidden;
+    margin: auto;
+}
+
+#admin-logo-preview-img {
+    max-width: 130px;
+    max-height: 30px;
+    width: auto;
+    height: auto;
+    display: block;
 }
 
 .admin-action {
